requirements.txt
requests
Flask
psycopg2-binary